+++ src/proto/misc2.pro
@@ -108,5 +108,4 @@ void parse_queued_messages(void);
 int mch_parse_cmd(char_u *cmd, int use_shcf, char ***argv, int *argc);
 int build_argv_from_string(char_u *cmd, char ***argv, int *argc);
 int build_argv_from_list(list_T *l, char ***argv, int *argc);
-int write_session_file(char_u *filename);
 /* vim: set ft=c : */
